Normalize fromContext with vdi page
