+++ @xen-orchestra/web/src/modules/storage-repository/components/list/StorageRepositoriesTable.vue
@@ -33,7 +33,6 @@ import {
   useXoSrCollection,
   type FrontXoSr,
 } from '@/modules/storage-repository/remote-resources/use-xo-sr-collection.ts'
-import { getSrScopeRouteQuery } from '@/modules/storage-repository/utils/xo-sr.util.ts'
 import { useXoRoutes } from '@/shared/remote-resources/use-xo-routes.ts'
 import VtsQueryBuilder from '@core/components/query-builder/VtsQueryBuilder.vue'
 import VtsRow from '@core/components/table/VtsRow.vue'
@@ -163,7 +162,7 @@ const { HeadCells, BodyCells } = useSrColumns({
           to: {
             name: '/sr/[id]',
             params: { id: sr.id },
-            query: getSrScopeRouteQuery(scope),
+            query: { from: scope.type, ...(scope.type === SR_SCOPE_TYPE.HOST && { host: scope.hostId }) },
           },
           icon: srStatusIcon.value,
           rightIcon: rightIcon.value,
